Authentication should now be secure ... and meant for web apps only.
The JWT is now stored in a secure HTTP-only cookie, as is a CSRF-token
in a cookie readable by Javascript (still secure). Protected endpoints
expect the header X-Csrf-Token to contain the value from the CSRF
cookie. Only if that token turns out to match too will the
authentication be OK.

The server is now SSL encrypted too.